Refactor: hightlightText shared/lib로  이동
